pyproject.toml
[tool.poetry]
name = "vuanem_ns"
version = "3.4"
description = ""
authors = ["HM <hieumdd@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.9"
google-cloud-bigquery = "^2.26.0"
requests = "^2.26.0"
functions-framework = "^2.2.1"
JayDeBeApi = "^1.2.3"
JPype1 = "^1.3.0"
Jinja2 = "^3.0.1"
SQLAlchemy = "^1.4.23"
psycopg2-binary = "^2.9.1"
google-cloud-tasks = "^2.5.1"
compose = "^1.2.8"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
pytest = "^6.2.5"
pytest-timeout = "^1.4.2"
mypy = "^0.942"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "-p no:warnings -s --durations=0"
testpaths = [
    "test",
]
timeout = 900
